# Improve plan summary performance