--- a/server/events/command/project_result_test.go
+++ b/server/events/command/project_result_test.go
@@ -225,3 +225,67 @@ func TestPlanSuccess_Summary(t *testing.T) {
 		})
 	}
 }
+
+var Summary string
+
+func BenchmarkPlanSuccess_Summary(b *testing.B) {
+	var s string
+
+	fixtures := map[string]string{
+		"changes": `
+					An execution plan has been generated and is shown below.
+					Resource actions are indicated with the following symbols:
+					  - destroy
+
+					Terraform will perform the following actions:
+
+					  - null_resource.hi[1]
+
+
+					Plan: 0 to add, 0 to change, 1 to destroy.`,
+		"no changes": `
+					An execution plan has been generated and is shown below.
+					Resource actions are indicated with the following symbols:
+
+					No changes. Infrastructure is up-to-date.`,
+		"changes outside Terraform": `
+					Note: Objects have changed outside of Terraform
+
+					Terraform detected the following changes made outside of Terraform since the
+					last "terraform apply":
+
+					No changes. Your infrastructure matches the configuration.`,
+		"changes and changes outside": `
+					Note: Objects have changed outside of Terraform
+
+					Terraform detected the following changes made outside of Terraform since the
+					last "terraform apply":
+
+					An execution plan has been generated and is shown below.
+					Resource actions are indicated with the following symbols:
+					  - destroy
+
+					Terraform will perform the following actions:
+
+					  - null_resource.hi[1]
+
+
+					Plan: 0 to add, 0 to change, 1 to destroy.`,
+		"empty summary, no matches": `No match, expect empty`,
+	}
+
+	for name, output := range fixtures {
+		p := &models.PlanSuccess{
+			TerraformOutput: output,
+		}
+
+		b.Run(name, func(b *testing.B) {
+			b.ReportAllocs()
+			for i := 0; i < b.N; i++ {
+				s = p.Summary()
+			}
+
+			Summary = s
+		})
+	}
+}