--- a/server/events/models/models.go
+++ b/server/events/models/models.go
@@ -366,20 +366,23 @@ type PlanSuccess struct {
 	HasDiverged bool
 }
 
+// Summary regexes
+var (
+	reChangesOutside = regexp.MustCompile(`Note: Objects have changed outside of Terraform`)
+	rePlanChanges    = regexp.MustCompile(`Plan: \d+ to add, \d+ to change, \d+ to destroy.`)
+	reNoChanges      = regexp.MustCompile(`No changes. (Infrastructure is up-to-date|Your infrastructure matches the configuration).`)
+)
+
 // Summary extracts one line summary of plan changes from TerraformOutput.
 func (p *PlanSuccess) Summary() string {
 	note := ""
-	r := regexp.MustCompile(`Note: Objects have changed outside of Terraform`)
-	if match := r.FindString(p.TerraformOutput); match != "" {
+	if match := reChangesOutside.FindString(p.TerraformOutput); match != "" {
 		note = fmt.Sprintf("\n**%s**\n", match)
 	}
-
-	r = regexp.MustCompile(`Plan: \d+ to add, \d+ to change, \d+ to destroy.`)
-	if match := r.FindString(p.TerraformOutput); match != "" {
+	if match := rePlanChanges.FindString(p.TerraformOutput); match != "" {
 		return note + match
 	}
-	r = regexp.MustCompile(`No changes. (Infrastructure is up-to-date|Your infrastructure matches the configuration).`)
-	return note + r.FindString(p.TerraformOutput)
+	return note + reNoChanges.FindString(p.TerraformOutput)
 }
 
 // DiffMarkdownFormattedTerraformOutput formats the Terraform output to match diff markdown format