@inkel 2023-01-02
I've noted that `PlanSuccess.DiffMarkdownFormattedTerraformOutput` also could be improved the same way, however, that one doesn't seem to have a specific test for it, so I haven't changed it yet.