@jamengual 2023-01-04
@inkel do you think you could add the improvements you mention in `PlanSuccess.DiffMarkdownFormattedTerraformOutput` in this PR?