@inkel 2023-01-04
@jamengual I sure can! Note that as I said above there are no tests for that specific method, but from what I can gather by reading the code it should be an easy to follow change and nothing should break.