@inkel 2023-01-04
I've found a test for it, it was in the `MarkdownRenderer` test suite. I'm going to add a benchmark there and the apply the improvements.